+++ RacingLeagueHub.Application/LeagueUsers/Persistence/ILeagueUserQueries.cs
@@ -4,7 +4,7 @@ namespace RacingLeagueHub.Application.LeagueUsers.Persistence;
 
 public interface ILeagueUserQueries
 { 
-    Task<List<LeagueUser>> GetAllLeagueRolesForUserAsync(long userId, CancellationToken cancellationToken = default);
-    Task<LeagueUser?> GetByLeagueAndUserAsync(long leagueId, long userId, CancellationToken cancellationToken = default);
-    Task<LeagueUser?> GetByLeagueAndUserAsync(string leagueSlug, long userId, CancellationToken cancellationToken = default);
+    Task<List<LeagueUser>> GetAllLeagueRolesForUserAsync(int userId, CancellationToken cancellationToken = default);
+    Task<LeagueUser?> GetByLeagueAndUserAsync(int leagueId, int userId, CancellationToken cancellationToken = default);
+    Task<LeagueUser?> GetByLeagueAndUserAsync(string leagueSlug, int userId, CancellationToken cancellationToken = default);
 }
